Added documentation, config for fee stuff, and moved functions around to appropriate places
