+++ src/main/resources/application.conf
@@ -20,18 +20,21 @@ logger.scala.slick=INFO
 
 lorre: {
   sleepIntervalInSeconds: 60
+  #Used to make sure Lorre records average fees every n iterations
+  feeUpdateInterval: 20
+  #Used to select how many fees should be averaged together
+  numberOfFeesAveraged: 1000
 }
 
 #Used for configuring await times of blocking operations
-dbAwaitTimeInSeconds: 60
+dbAwaitTimeInSeconds: 120
 
-#Used to make sure Lorre records average fees every n iterations
-feeUpdateInterval: 20
 
 platforms: {
   tezos: {
     zeronet : {
       node: {
+        protocol: "https"
         hostname: "nautilus.cryptonomic.tech",
         port: 8732
         pathPrefix: "tezos/zeronet/"
