+++ src/main/scala/tech/cryptonomic/conseil/tezos/TezosNodeInterface.scala
@@ -49,10 +49,11 @@ object TezosNodeInterface extends TezosRPCInterface with LazyLogging {
   @Override
   def runGetQuery(network: String, command: String, payload: Option[String]= None): Try[String] = {
     Try{
+      val protocol = conf.getString(s"platforms.tezos.$network.node.protocol")
       val hostname = conf.getString(s"platforms.tezos.$network.node.hostname")
       val port = conf.getInt(s"platforms.tezos.$network.node.port")
       val pathPrefix = conf.getString(s"platforms.tezos.$network.node.pathPrefix")
-      val url = s"http://$hostname:$port/${pathPrefix}chains/main/$command"
+      val url = s"$protocol://$hostname:$port/${pathPrefix}chains/main/$command"
       logger.info(s"Querying URL $url for platform Tezos and network $network with payload $payload")
       val responseFuture: Future[HttpResponse] =
         Http(system).singleRequest(
@@ -72,10 +73,11 @@ object TezosNodeInterface extends TezosRPCInterface with LazyLogging {
   @Override
   def runPostQuery(network: String, command: String, payload: Option[String]= None): Try[String] = {
     Try{
+      val protocol = conf.getString(s"platforms.tezos.$network.node.protocol")
       val hostname = conf.getString(s"platforms.tezos.$network.node.hostname")
       val port = conf.getInt(s"platforms.tezos.$network.node.port")
       val pathPrefix = conf.getString(s"platforms.tezos.$network.node.pathPrefix")
-      val url = s"http://$hostname:$port/${pathPrefix}chains/main/$command"
+      val url = s"$protocol://$hostname:$port/${pathPrefix}chains/main/$command"
       logger.info(s"Querying URL $url for platform Tezos and network $network with payload $payload")
       val postedData = payload match {
         case None => """{}"""
